fixed compilation failure
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@14081 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ src/com/l7tech/common/io/failover/RoundRobinFailoverStrategy.java
@@ -48,7 +48,7 @@ public class RoundRobinFailoverStrategy extends AbstractFailoverStrategy {
                 Map.Entry entry = (Map.Entry)i.next();
                 Object server = entry.getKey();
                 Long probeWhen = (Long)entry.getValue();
-                if (probeWhen >= now) {
+                if (probeWhen.longValue() >= now) {
                     // Probe this server; update time, move to end of list, and return it
                     i.remove();
                     down.put(server, new Long(now + probeTime));
